amazon/order-close/requirements.txt
python-amazon-sp-api
boto3
requests
